// src/types.ts
export type _USER_DATA = {
  id: number;
  EnrollmentNo: String;
  Name: String;
  NameH: String;
  EmailID: String;
  IITREmailID: String;
  ProgramID: String;
  Mobileno: String;
  AadharCardNo: String;
  SemesterID: String;
  StructureID: String;
  DepartmentID: String;
  Fathersname: String;
  FatherMobileNo: String;
  FatherIncome: String;
  FatherOccupation: String;
  MotherName: String;
  MotherMobileNo: String;
  MotherIncome: String;
  MotherOccupation: String;
  Nationality: String;
  StateofDomicile: String;
  Religion: String;
  BelongstoMinority: String;
  NameofMinority: String;
  DateofBirth: String;
  MaritalStatus: String;
  Category: String;
  SubCategory: String;
  IsPrep: String;
  Gender: String;
  PersonwithDisability: String;
  DisabilityType: String;
  LocalGaurdian: String;
  LocalGaurdianPhone: String;
  LocalGaurdianAddress: String;
  PermanentAddress: String;
  State: String;
  City: String;
  Pincode: String;
  Pcountry: String;
  ContactNo: String;
  PEmailid: String;
  ExamQualified: String;
  Score_Rank: String;
  Categoryrank: String;
  ValidityUpto: String;
  FellowShip: String;
  StatusPHD: String;
  TitleofTheses: String;
  SuperviserDetail: String;
  Bhawan: String;
  RoomNo: String;
  BloodGroup: String;
  Remark: String;
  StStatus: String;
  StartSession: String;
  SemNo: String;
  StPic: String;
  Batch: String;
  SubBatch: String;
  StSig: String;
  Awd: String;
  P_Email: String;
  RegCode: String;
  ProgramAlphaCode: String;
  DepartmentAlphaCode: String;
  Expr1: String;
  ProgramName: String;
  DepartmentName: String;
  Prgname: String;
};